chore: move utils file
