+++ ui/src/pages/session-details/playback/utils.tsx
